testing: organize fixtures and test node_fetch_host_keys
